[BOOK-194] fix: resolve conflicts
